fix: module rename
