+++ js/main.js
@@ -1,6 +1,6 @@
 import './functions/get-random-integer.js';
 import './functions/get-random-array-element.js';
-import './functions/unique-number-generator.js';
+import './functions/create-unique-numbers-generator.js';
 import { createPhotos } from './data.js';
 
 /* eslint-disable no-console */
